C:/Program Files/Git/play behaviour changed, /play structure updated
