+++ src/commands/basic/loop.js
@@ -10,7 +10,7 @@ module.exports = {
             const queue = useQueue(interaction.guild.id);
             queue.setRepeatMode(1);
 
-            await interaction.reply('Track was looped!')
+            await interaction.reply('Track was looped')
         } catch(e) {
             interaction.reply('loop error :', e);
         }
