+++ src/commands/basic/next.js
@@ -10,7 +10,7 @@ module.exports = {
             const queue = useQueue(interaction.guild.id);
             queue.node.skip();
 
-            await interaction.reply('Start playing next track!')
+            await interaction.reply('Start playing next track')
         } catch(e) {
             return interaction.reply('next error', e)
         }
