+++ src/commands/basic/pause.js
@@ -10,7 +10,7 @@ module.exports = {
             const queue = useQueue(interaction.guild.id);
             queue.node.setPaused(true);
 
-            await interaction.reply("Track was paused!");
+            await interaction.reply("Track was paused");
         } catch (e) {
             interaction.reply("pause error ", e)
         }
