+++ src/commands/basic/ping.js
@@ -3,7 +3,7 @@ const { SlashCommandBuilder } = require('discord.js');
 module.exports = {
 	data: new SlashCommandBuilder()
 		.setName('ping')
-		.setDescription('Replies with Pong!'),
+		.setDescription('Replies with Pong'),
 	async execute({ client, interaction }) {
 		await interaction.reply('Pong!');
 	},
