+++ src/commands/basic/resume.js
@@ -10,10 +10,10 @@ module.exports = {
             const queue = useQueue(interaction.guild.id);
             queue.node.setPaused(false);
 
-            await interaction.reply("Track was resumed!");
+            await interaction.reply("Track was resumed");
 
         } catch (e) {
-            return interaction.reply("resume error ", e)
+            return interaction.reply("Resume error: ", e)
         }
     }
 }
