+++ src/commands/basic/unloop.js
@@ -10,9 +10,9 @@ module.exports = {
             const queue = useQueue(interaction.guild.id);
             queue.setRepeatMode(0);
 
-            await interaction.reply('Track was unlooped !')
+            await interaction.reply('Track was unlooped')
         } catch (e) {
-            return interaction.reply("unloop error ", e)
+            return interaction.reply("Unloop error", e)
         }
 
     }
